changed proposed solution to the out of bound error
instead of adding an if else statement, I moved the placement of assigning a value on the source attribute. If the source attribute is missing from the metadata, and it's the first value that's added to the `data_info` list, the rest of the list can't be populated, and it returns the "out of bound" error.